+++ t-unzipper/src/main/java/eu/unifiedviews/plugins/transformer/unzipper/UnZipperVaadinDialog.java
@@ -8,7 +8,7 @@ import eu.unifiedviews.helpers.dpu.vaadin.dialog.AbstractDialog;
 
 public class UnZipperVaadinDialog extends AbstractDialog<UnZipperConfig_V1> {
 
-    private CheckBox checkNotPrefix;
+    private CheckBox checkDuplicityAvoid;
     
     public UnZipperVaadinDialog() {
         super(UnZipper.class);
@@ -22,9 +22,9 @@ public class UnZipperVaadinDialog extends AbstractDialog<UnZipperConfig_V1> {
         mainLayout.setWidth("100%");
         mainLayout.setHeight("-1px");
 
-        checkNotPrefix = new CheckBox(ctx.tr("unzipper.dialog.unzip.noprefix"));
-        checkNotPrefix.setDescription(ctx.tr("unzipper.dialog.unzip.noprefix.description"));
-        mainLayout.addComponent(checkNotPrefix);
+        checkDuplicityAvoid = new CheckBox(ctx.tr("unzipper.dialog.unzip.duplicity.avoid"));
+        checkDuplicityAvoid.setDescription(ctx.tr("unzipper.dialog.unzip.duplicity.avoid.description"));
+        mainLayout.addComponent(checkDuplicityAvoid);
 
         setCompositionRoot(mainLayout);
         
@@ -32,13 +32,17 @@ public class UnZipperVaadinDialog extends AbstractDialog<UnZipperConfig_V1> {
 
     @Override
     protected void setConfiguration(UnZipperConfig_V1 c) throws DPUConfigException {
-        checkNotPrefix.setValue(c.isNotPrefixed());
+        // We have changed the description of filed in user GUI to negation of old description, so we have to negate the result
+        // and we do not want to change config class
+        checkDuplicityAvoid.setValue(!c.isNotPrefixed());
     }
 
     @Override
     protected UnZipperConfig_V1 getConfiguration() throws DPUConfigException {
         final UnZipperConfig_V1 cnf = new UnZipperConfig_V1();
-        cnf.setNotPrefixed(checkNotPrefix.getValue() == null ? false : checkNotPrefix.getValue());
+        // We have changed the description of filed in user GUI to negation of old description, so we have to negate the result
+        // and we do not want to change config class
+        cnf.setNotPrefixed(checkDuplicityAvoid.getValue() == null ? true : !checkDuplicityAvoid.getValue());
         return cnf;
     }
 
@@ -46,12 +50,12 @@ public class UnZipperVaadinDialog extends AbstractDialog<UnZipperConfig_V1> {
     public String getDescription() {
         StringBuilder desc = new StringBuilder();
 
-        if (checkNotPrefix.getValue() == true) {
+        if (checkDuplicityAvoid.getValue() == true) {
             // If true then we do not use prefixes.
-            desc.append(ctx.tr("unzipper.dialog.unzip.notprefixed"));
+            desc.append(ctx.tr("unzipper.dialog.unzip.duplicity.avoid.on"));
         } else {
             // If false prefix is not used.
-            desc.append(ctx.tr("unzipper.dialog.unzip.prefixed"));
+            desc.append(ctx.tr("unzipper.dialog.unzip.duplicity.avoid.off"));
         }
 
         return desc.toString();